Update SPOC layout for long titles (#54)
* Adding SPOC specific files
* Moving SPOC to its own component
* Adding proper SPOC header padding
* Adjusting design from feedback on SPOC
