+++ src/containers/save/recommendations/list/spoc.js
@@ -3,8 +3,8 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames/bind'
 import * as Icon from '../../../../components/icons'
-import { domainForUrl } from '../../../../common/utilities'
 import { localize } from '../../../../common/_locales/locales'
+import SpocHeader from './spoc.header'
 
 const cx = classNames.bind(styles)
 const copy = {
@@ -12,7 +12,7 @@ const copy = {
   saving: localize('recommendations', 'saving'),
   saved: localize('recommendations', 'saved')
 }
-export default class RecommendationItem extends Component {
+export default class RecommendationSpoc extends Component {
   get imageStyle() {
     return {
       backgroundImage: 'url("' + this.props.item.image + '")',
@@ -25,8 +25,21 @@ export default class RecommendationItem extends Component {
     return copy[this.props.item.status]
   }
 
+  get spocContext() {
+    const item = this.props.item
+    return {
+      cxt_impression_id: item.impression_id,
+      cxt_view: 'extension_ad',
+      cxt_feed_item: item.feed_item_id,
+      cxt_index: item.sort_id,
+      cxt_post_id: item.post_id
+    }
+  }
+
   onClick = () => {
     const item = this.props.item
+
+    this.props.spocClick({ context: this.spocContext })
 
     this.props.openRecommendation({
       tabId: this.props.tabId,
@@ -41,11 +54,6 @@ export default class RecommendationItem extends Component {
   render() {
     let item = this.props.item
 
-    let itemContainerClass = cx({
-      itemContainer: true,
-      isSpoc: item.isSpoc
-    })
-
     let recommendationItemClass = cx({
       item: true,
       hasImage: item.has_image
@@ -59,7 +67,20 @@ export default class RecommendationItem extends Component {
 
     return (
       <li style={this.props.motionStyle} className={recommendationItemClass}>
-        <div className={itemContainerClass}>
+        <SpocHeader
+          tabId={this.props.tabId}
+          itemId={item.id.toString()}
+          sponsorurl={item.url}
+          sponsor={item.sponsor}
+          avatar={item.avatar}
+          spocContext={this.spocContext}
+          spocImpression={this.props.spocImpression}
+          spocView={this.props.spocView}
+          spocClick={this.props.spocClick}
+          spocRemove={this.props.spocRemove}
+        />
+
+        <div className={styles.spocContainer}>
           {item.has_image && (
             <div className={styles.image} style={this.imageStyle} />
           )}
@@ -75,7 +96,7 @@ export default class RecommendationItem extends Component {
             </a>
           </div>
 
-          <div className={styles.source}>{domainForUrl(item.display_url)}</div>
+          <div className={styles.source}>{item.display_url}</div>
 
           <div className={styles.actions}>
             <button
@@ -99,7 +120,7 @@ export default class RecommendationItem extends Component {
   }
 }
 
-RecommendationItem.propTypes = {
+RecommendationSpoc.propTypes = {
   item: PropTypes.object,
   motionStyle: PropTypes.object,
   saveRec: PropTypes.func
